theme: Flatten the theme directory instead of nesting a copied version of Solbit
Before this change, it was not easy to see where changes to the theme
should be made. The commands were also never updated to the new
structure, leading contributors to believe that the old theme repos
were still in use. The old Makefile contents should probably be ported
to the new Taskfile, but that can be its own PR.

This should hopefully make it easier to navigate the theme. Or, at least
make it much less annoying to work on.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>