editions: Bring back the 2x2 grid
This brings back the classic 2x2 grid that the existing website has.
There are still a few remaining issues, mainly that I don't know where
the checksum arrow went. Also dark mode needs a lot of help.

This also adds a section at the bottom of the page on ISO verification.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>