index: Redesign the home page
This puts a lot more information on the home page instead of having it
spread in several places across the website. It still needs images for
each feature section, and there is a desire to have background colors
for some of them. The latter point I think will require overriding the
  `hextra-home` layout, but we could do it.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>