shortcodes: Create a reusable button component
Sometimes, we want an actual button that isn't a card. Our button
supports two forms: normal and primary. This is useful when we have two
buttons next to each other, and we want to highlight one of them.

This also sets up the primary color for the theme. This color is set to
the lighter blue in the Solus logo. So far, only the buttons use it, I
think. Hextra may use it for other components, though, such as blog
elements; we'll find out when we get there.

Lastly, the Help Center and Download buttons on the home page have been
converted to our new buttons.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>