navbar: Switch to the logo-only SVG for the logo
The previous logo-with-text made the navbar very cluttered on mobile
views. This changes it to display the logo and text separately, enabling
us to use CSS to hide the title on mobile.

To further de-clutter the mobile navbar, items with icons have been
moved to the sidebar on mobile, leaving just the logo and menu button.

This also has a fix for the sidebar footer in dark mode thrown in for
good measure.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>