team: Add a page giving an overview of our team
There is a page for this on the Help Center already, but I think it
makes sense to have it on the main website. There is a larger
conversation we could have about whether the entire Organization section
of the Help Center should be moved to the main website, but that can
come later.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>